+++ back/src/forms/resolvers/mutations/markAsSealed.ts
@@ -4,7 +4,7 @@ import { MutationResolvers } from "../../../generated/graphql/types";
 import { getFormOrFormNotFound } from "../../database";
 import { expandFormFromDb } from "../../form-converter";
 import { checkCanMarkAsSealed } from "../../permissions";
-import { checkCanBeSealed } from "../../validation";
+import { checkCanBeSealed, checkCompaniesType } from "../../validation";
 import transitionForm from "../../workflow/transitionForm";
 import { EventType } from "../../workflow/types";
 
@@ -19,6 +19,9 @@ const markAsSealedResolver: MutationResolvers["markAsSealed"] = async (
 
   // validate form data
   await checkCanBeSealed(form);
+
+  await checkCompaniesType(form);
+
   const sealedForm = await transitionForm(user, form, {
     type: EventType.MarkAsSealed
   });